requirements.txt
torch==2.1.0
transformers==4.34.1
tabulate==0.9.0
Django~=4.2.7
djangorestframework~=3.14.0
numpy~=1.26.2
Pillow~=10.0.1
paddleocr~=2.6.1.3
